feat: update
1. cli completed
2. remove gui and add gui later

+++ model/dao/grpcutils.cpp
@@ -42,10 +42,10 @@ void setupCtx(grpc::ClientContext &ctx)
                      std::chrono::milliseconds(STQ_CLIENT_TIMEOUT * 1000));
 }
 
-void buildErrMsg(const char *file, i8 line, grpc::Status &status)
+void buildErrMsg(const char *file, i32 line, grpc::Status &status)
 {
     spdlog::error("{}:{} gRPC error code {}: {}", file, line,
-                  static_cast<i8>(status.error_code()), status.error_message());
+                  static_cast<i32>(status.error_code()), status.error_message());
 }
 
 } // end namespace GRPCUtils
